requirements.txt
flask==2.1.0
beautifulsoup4==4.10.0
pandas==1.3.3
numpy==1.21.2
PyMuPDF==1.18.14
tabula-py==2.3.0
mysql-connector-python==8.0.27
